+++ android/app/src/main/java/org/libsdl/app/SDLActivity.java
@@ -1336,6 +1336,7 @@ public class SDLActivity extends Activity implements View.OnSystemUiVisibilityCh
     public static native void setTutorialLanguage(boolean german);
     public static native boolean isTutorialVisible();
     public static native void requestTouchPresetUpdateNotice(int version);
+    public static native void requestTouchOverlayFeatureNotice(int version);
 
     public static boolean handleKeyEvent(View v, int keyCode, KeyEvent event, InputConnection ic) {
         int deviceId = event.getDeviceId();